Separated methods for EmployeeHierarchy into other controller.
