+++ src/main/java/dev/austinbarnes/retailinventorymanagement/employee/controller/EmployeeController.java
@@ -1,12 +1,8 @@
 package dev.austinbarnes.retailinventorymanagement.employee.controller;
 
 import dev.austinbarnes.retailinventorymanagement.common.ApiResponseDto;
-import dev.austinbarnes.retailinventorymanagement.employee.dto.employee.EmployeeHierarchyRequestDTO;
-import dev.austinbarnes.retailinventorymanagement.employee.dto.employee.EmployeeHierarchyResponseDTO;
 import dev.austinbarnes.retailinventorymanagement.employee.dto.employee.EmployeeRequestDTO;
 import dev.austinbarnes.retailinventorymanagement.employee.dto.employee.EmployeeResponseDTO;
-import dev.austinbarnes.retailinventorymanagement.employee.repo.EmployeeRepository;
-import dev.austinbarnes.retailinventorymanagement.employee.service.EmployeeHierarchyService;
 import dev.austinbarnes.retailinventorymanagement.employee.service.EmployeeService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -18,19 +14,25 @@ import java.util.List;
 import java.util.UUID;
 
 
-
 @RestController
 @RequestMapping("api/v1/employees")
 
 @Slf4j
 @RequiredArgsConstructor
 /*
- * Controller for Employee Entity as well as EmployeeHierarchy
+  EmployeeController handles all employee-related operations.
+  It provides endpoints to create, update, delete, and retrieve employee information.
  */
 public class EmployeeController {
     private final EmployeeService employeeService;
-    private final EmployeeHierarchyService hierarchyService;
 
+    /**
+     * Creates a new employee.
+     *
+     * @param request   The employee request DTO containing employee details.
+     * @param managerId Optional manager ID for the new employee.
+     * @return ResponseEntity with the created employee details.
+     */
     @PostMapping
     public ResponseEntity<ApiResponseDto<EmployeeResponseDTO>> createEmployee(
             @RequestBody @Valid EmployeeRequestDTO request,
@@ -45,62 +47,52 @@ public class EmployeeController {
                 employeeService.createEmployee(request);
     }
 
+    /**
+     * Retrieves all employees.
+     *
+     * @return ResponseEntity with a list of all employees.
+     */
     @GetMapping
     public ResponseEntity<ApiResponseDto<List<EmployeeResponseDTO>>> getAllEmployees() {
         log.info("Get all employees");
         return employeeService.getAllEmployees();
     }
 
+    /**
+     * Retrieves an employee by ID.
+     *
+     * @param employeeId The ID of the employee to retrieve.
+     * @return ResponseEntity with the employee details.
+     */
     @GetMapping("/{employeeId}")
     public ResponseEntity<ApiResponseDto<EmployeeResponseDTO>> getEmployeeById(@PathVariable UUID employeeId) {
         log.info("Get employee by id: {}", employeeId);
         return employeeService.getEmployeeById(employeeId);
     }
 
+    /**
+     * Updates an existing employee.
+     *
+     * @param employeeId The ID of the employee to update.
+     * @param request    The employee request DTO containing updated employee details.
+     * @return ResponseEntity with the updated employee details.
+     */
     @PutMapping("/{employeeId}")
     public ResponseEntity<ApiResponseDto<EmployeeResponseDTO>> updateEmployee(@PathVariable UUID employeeId, @RequestBody @Valid EmployeeRequestDTO request) {
         log.info("Update employeeId: {} with request: {}", employeeId, request);
         return employeeService.updateEmployee(employeeId, request);
     }
 
+    /**
+     * Deletes an employee by ID.
+     *
+     * @param employeeId The ID of the employee to delete.
+     * @return ResponseEntity with no content.
+     */
     @DeleteMapping("/{employeeId}")
     public ResponseEntity<ApiResponseDto<Void>> deleteEmployee(@PathVariable UUID employeeId) {
         log.info("Delete employee {}", employeeId);
         return employeeService.deleteEmployee(employeeId);
-    }
-
-    @PostMapping("/hierarchy")
-    public ResponseEntity<ApiResponseDto<EmployeeHierarchyResponseDTO>> createEmployeeHierarchy(
-            @RequestBody @Valid EmployeeHierarchyRequestDTO request
-    ) {
-        log.info("Create employee hierarchy request: {}", request);
-        return hierarchyService.createEmployeeHierarchy(request);
-    }
-
-    @PutMapping("/hierarchy")
-    public ResponseEntity<ApiResponseDto<EmployeeHierarchyResponseDTO>> updateEmployeeHierarchy(
-            @RequestBody @Valid EmployeeHierarchyRequestDTO request
-    ) {
-        log.info("Update employee hierarchy request: {}", request);
-        return hierarchyService.updateHierarchy(request);
-    }
-
-    @GetMapping("/hierarchy/{employeeId}")
-    public ResponseEntity<ApiResponseDto<EmployeeHierarchyResponseDTO>> getEmployeeHierarchy(@PathVariable UUID employeeId) {
-        log.info("Get employee hierarchy request: {}", employeeId);
-        return hierarchyService.getEmployeeHierarchy(employeeId);
-    }
-
-    @GetMapping("/hierarchy")
-    public ResponseEntity<ApiResponseDto<List<EmployeeHierarchyResponseDTO>>> getEmployeeHierarchyAll() {
-        log.info("Get employee hierarchy all");
-        return hierarchyService.getEmployeeHierarchyAll();
-    }
-
-    @DeleteMapping("/hierarchy")
-    public ResponseEntity<ApiResponseDto<Void>> deleteEmployeeHierarchyRelationship(UUID employeeHierarchyId) {
-        log.info("Delete hierarchy relationship {}", employeeHierarchyId);
-        return hierarchyService.deleteEmployeeHierarchyRelationship(employeeHierarchyId);
     }
 }
