+++ src/main/java/dev/austinbarnes/retailinventorymanagement/employee/repo/EmployeeHierarchyRepository.java
@@ -11,5 +11,4 @@ import java.util.UUID;
 @Repository
 public interface EmployeeHierarchyRepository extends JpaRepository<EmployeeHierarchy, UUID> {
     Optional<EmployeeHierarchy> findByEmployeeId(UUID employeeId);
-    List<EmployeeHierarchy> findAllByEmployeeId(UUID employeeId);
 }
